Move CUBE\ to it's own module
